Changes done after Code  review - Changed areaid/dealerid datatype to uint32 - Added comments

+++ BikewaleOpr.Entity/GeoLocationEntity.cs
@@ -8,7 +8,7 @@ namespace BikewaleOpr.Entities
     /// </summary>
     public class GeoLocationEntity
     {
-        public UInt16 Id { get; set; }
+        public UInt32 Id { get; set; }
         public double Latitude { get; set; }
         public double Longitude { get; set; }
     }
